+++ website/kanikervanaf/templates/kanikervanaf/base.html
@@ -83,7 +83,7 @@
                             <a class="nav-link" href="{% url "subscriptions:overview" %}">1. Selecteer abonnementen</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="{% url "users:enter" %}">2. Voer uw gegevens in</a>
+                            <a class="nav-link" href="{% url "subscriptions:enter" %}">2. Voer uw gegevens in</a>
                         </li>
                         <li class="nav-item">
                             <a class="nav-link" href="{% url "subscriptions:summary" %}">3. Versturen</a>
