+++ website/subscriptions/sitemaps.py
@@ -160,3 +160,27 @@ class SubscriptionDetailsSitemap(Sitemap):
         :return: the relative url to the objects location on this site
         """
         return reverse("subscriptions:details", kwargs={"subscription": obj})
+
+
+class EnterUserSitemap(Sitemap):
+    """Sitemap for EnterUserinformationView."""
+
+    changefreq = "never"
+    priority = 0.6
+
+    def items(self):
+        """
+        Get a list of items corresponding to this sitemap.
+
+        :return: a list of urls
+        """
+        return ["subscriptions:enter"]
+
+    def location(self, obj):
+        """
+        Get the location of an object for this sitemap.
+
+        :param obj: the object to get the location for
+        :return: the relative url to the objects location on this site
+        """
+        return reverse(obj)
