+++ website/subscriptions/templates/subscriptions/enter_information.html
@@ -29,5 +29,5 @@
 
 {% block js %}
     <script src="{% static 'subscriptions/js/general.js' %}"></script>
-    <script src="{% static 'users/js/details.js' %}"></script>
+    <script src="{% static 'subscriptions/js/details.js' %}"></script>
 {% endblock %}
