+++ website/subscriptions/templates/subscriptions/subscription_category.html
@@ -46,5 +46,5 @@
     </div>
 {% endblock %}
 {% block sidebar %}
-    {% include "subscriptions/subscription_list.html" with next_page="users:enter" button_enabled="subscriptions" %}
+    {% include "subscriptions/subscription_list.html" with next_page="subscriptions:enter" button_enabled="subscriptions" %}
 {% endblock %}
