+++ website/subscriptions/templates/subscriptions/subscription_category_page.html
@@ -41,5 +41,5 @@
     </div>
 {% endblock %}
 {% block sidebar %}
-    {% include "subscriptions/subscription_list.html" with next_page="users:enter" button_enabled="subscriptions" scroller=True %}
+    {% include "subscriptions/subscription_list.html" with next_page="subscriptions:enter" button_enabled="subscriptions" scroller=True %}
 {% endblock %}
