+++ website/subscriptions/templates/subscriptions/verification_send.html
@@ -18,7 +18,7 @@
             <h1 class="text-center mb-5"><strong>Verificatie verzenden mislukt!</strong></h1>
             <p class="text-center">Het verzenden van de verificatiemail is mislukt, heb je het juist email-adres ingevuld?
                 <br>
-                <a class="btn btn-primary mt-3" href="{% url 'users:enter' %}"><i class="fas fa-angle-double-left"></i> Terug naar mijn gegevens</a>
+                <a class="btn btn-primary mt-3" href="{% url 'subscriptions:enter' %}"><i class="fas fa-angle-double-left"></i> Terug naar mijn gegevens</a>
             </p>
         {% endif %}
     </div>
