+++ website/users/views.py
@@ -11,54 +11,12 @@ from .forms import (
     PasswordResetForm,
     PasswordUpdateForm,
     UserUpdateForm,
-    EnterUserInformationForm,
     EmailUpdateForm,
 )
 from .models import User, PasswordReset, EmailUpdate
 from .services import generate_password_reset, send_reset_password, send_update_email
 import json
 from urllib.parse import unquote_plus
-
-
-class BasicUserInformation(TemplateView):
-    """View for entering user information."""
-
-    cookie_name = "subscription_details"
-
-    template_name = "users/enter_information.html"
-
-    def _is_cookie_empty(self, cookie):
-        if cookie is not None:
-            try:
-                loaded_data = json.loads(unquote_plus(cookie))
-            except json.JSONDecodeError:
-                return True
-            if type(loaded_data) == dict:
-                for key in loaded_data.keys():
-                    if bool(loaded_data[key]):
-                        return False
-        return True
-
-    def get(self, request, **kwargs):
-        """
-        GET request for user information view.
-
-        :param request: the request
-        :param kwargs: keyword arguments
-        :return: render of the enter_information page
-        """
-        if request.user.is_authenticated and self._is_cookie_empty(
-            request.COOKIES.get("subscription_details", None)
-        ):
-            return render(
-                request,
-                self.template_name,
-                {"form": EnterUserInformationForm(user=request.user)},
-            )
-        else:
-            return render(
-                request, self.template_name, {"form": EnterUserInformationForm()}
-            )
 
 
 class LoginView(TemplateView):
